requirement.txt
torch
torchvision
pyyaml
tqdm
numpy
scipy
matplotlib
pillow
thop
mmdet==2.28.2
mmcv-full==1.7.2